Added mvc for user authentication
