+++ models/index.js
@@ -1,4 +1,4 @@
 module.exports = {
   User: require('./User'),
-  Post: require('./Post');
+  Post: require('./Post')
 };
