+++ server.js
@@ -1,13 +1,12 @@
 const express = require('express');
 const path = require('path');
 const bodyParser = require('body-parser');
-const jwt = require('jsonwebtoken');
-const exjwt = require('express-jwt');
 const mongoose = require('mongoose');
 const morgan = require('morgan'); // used to see requests
 const app = express();
 const db = require('./models');
 const PORT = process.env.PORT || 3001;
+const routes = require("./routes");
 
 // Setting CORS so that any website can
 // Access our API
@@ -26,56 +25,12 @@ app.use(bodyParser.urlencoded({ extended: true }));
 
 mongoose.connect(process.env.MONGODB_URI || 'mongodb://localhost/whistlrDB');
 
-// Init the express-jwt middleware
-const isAuthenticated = exjwt({
-  secret: 'all sorts of code up in here'
-});
-
-
-// LOGIN ROUTE
-app.post('/api/login', (req, res) => {
-  db.User.findOne({
-    email: req.body.email
-  }).then(user => {
-    user.verifyPassword(req.body.password, (err, isMatch) => {
-      if(isMatch && !err) {
-        let token = jwt.sign({ id: user._id, email: user.email }, 'all sorts of code up in here', { expiresIn: 129600 }); // Sigining the token
-        res.json({success: true, message: "Token Issued!", token: token, user: user});
-      } else {
-        res.status(401).json({success: false, message: "Authentication failed. Wrong password."});
-      }
-    });
-  }).catch(err => res.status(404).json({success: false, message: "User not found", error: err}));
-});
-
-// SIGNUP ROUTE
-app.post('/api/signup', (req, res) => {
-  db.User.create(req.body)
-    .then(data => res.json(data))
-    .catch(err => res.status(400).json(err));
-});
-
-// Any route with isAuthenticated is protected and you need a valid token
-// to access
-app.get('/api/user/:id', isAuthenticated, (req, res) => {
-  db.User.findById(req.params.id).then(data => {
-    if(data) {
-      res.json(data);
-    } else {
-      res.status(404).send({success: false, message: 'No user found'});
-    }
-  }).catch(err => res.status(400).send(err));
-});
-
 // Serve up static assets (usually on heroku)
 if (process.env.NODE_ENV === "production") {
   app.use(express.static("client/build"));
 }
 
-
-app.get('/', isAuthenticated /* Using the express jwt MW here */, (req, res) => {
-  res.send('You are authenticated'); //Sending some response when authenticated
-});
+app.use(routes);
 
 // Error handling
 app.use(function (err, req, res, next) {
@@ -85,12 +40,6 @@ app.use(function (err, req, res, next) {
   else {
     next(err);
   }
-});
-
-// Send every request to the React app
-// Define any API routes before this runs
-app.get("*", function(req, res) {
-  res.sendFile(path.join(__dirname, "./client/build/index.html"));
 });
 
 app.listen(PORT, function() {
